pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tsmp-oop"
version = "0.1.0"
description = "OOP time-series modeling platform with constraint optimization and context adaptation"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.0.0",
  "pyarrow>=12.0.0",
  "numpy>=1.24.0",
  "sqlalchemy>=2.0.0",
  "psycopg2-binary>=2.9.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0",
  "mlflow>=2.8.0",
  "psutil>=5.9.0",
  "pynvml>=11.5.0",
  "scikit-learn>=1.3.0",
  "fastapi>=0.104.0",
  "uvicorn>=0.24.0"
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "ruff>=0.1.0", "mypy>=1.6.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -s"
